requirements.txt
rembg
opencv-python
Pillow
customtkinter
tkinter